Solve equations where the LHS or RHS is a single constant
And where the opposite side normalises to a constant. Previously
when a wanted constraint had a constant LHS or RHS, we would give
up. Now we check if the normalised version of the opposite side
differs from the original version of the opposite side. When they
differ, we see if the LHS and RHS are equal, or if there exists
a unifier.