Use coercion holes of existing wantedw when possible
Before, we did not generate new [W]anteds, when we had already
seen a [W]anted constraints with the same predicate type.
Sometimes, however, GHC does not reuse evidence for a wanted
constraint with an equal predicate type.
Consequently, we could only solve a certain constraint once,
even though GHC wanted us to solve it twice.

What we do now, is only created new [W]anted contraints, if
GHC does not have a [W]anted constraint in scope with the same
predicate type. For the other desired constraints, we use
the coercion holes of an existing [W]anted constraint with the
same predicate type.