Always generate new [W]anted constraints
Before, we did not generate new [W]anteds, when we had already
seen a [W]anted constraint with the same predicate type.
Sometimes, however, GHC does not reuse evidence for a wanted
constraint with an equal predicate type.
Consequently, we could only solve a certain constraint once,
even though GHC wanted us to solve it twice.

The reason that we used to have this "have-we-seen-this-wanted"
logic in, is that in GHC 7.10 GHC could go into an endless
constraint solving loop if a plugin would keep generating new
wanteds. However, as of GHC 8, GHC stops with an error that it
has done too many solver iterations.