Solving for `a - b` implies `b <= a`
When we find a solution for a contraint involving a subtraction
than we must emit an inequality contraint saying that the
RHS of the subtraction is smaller than the LHS. If we don't
emit this constraint, we get to prove:

```
(n - 1) + 1 ~ n
```

even for any `n`, even though for `n~0`, `n-1` is not a natural
number. We must hence emit an `1 <= n` constraint when giving
a proof for the above equation.

Fixes #14