Don't solve constraints when there's no most general solution
Do not solve `x + y ~ a + b` by emitting `x ~ a` and `y ~ b`
constraints, left to handle by the rest of the constraint solving
pipeline; because another valid solution might be emitting
`x ~ b` and `y ~ a` constraints. Picking a solution that is
not the most general can lead so unexpected situations.